TEST: fix test in test_ephem.py
Due to a rounding error in sinc_interp() of boinor._math.interpolate
a test in test_ephem.py failed. Two values should have been equal,
but the test needed a assert_coordinates_allclose() with a very small
tol to pass.

In case of very small values of y_u[] in sinc_interp(), these values
are now set to zero. So basically taking care of rounding errors
in a central function.